pyproject.toml
[project]
name = "image-stitching-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.12",
    "numpy>=2.2.6",
    "opencv-python>=4.11.0.86",
    "pillow>=11.2.1",
    "python-multipart>=0.0.20",
    "uvicorn>=0.34.2",
]
